+++ gradle.properties
@@ -9,7 +9,7 @@ yarn_mappings=1.21+build.9
 loader_version=0.16.14
 
 # Mod Properties
-mod_version=1.0.0
+mod_version=1.1.0
 maven_group=me.onethecrazy
 archives_base_name=tableofuncrafting
