+++ src/main/java/me/onethecrazy/TableOfUncraftingDataGenerator.java
@@ -1,6 +1,5 @@
 package me.onethecrazy;
 
-import me.RecipeGenerator;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
